remove unnecessary os-specific implementations now that we have abstracted out the filesystem
